+++ include/ingamelogic.h
@@ -22,8 +22,8 @@ public:
     void handlenetforlevel(uint8_t* data,size_t length,int result);
     std::shared_ptr<GameState> getgamestate();
     void creategamestate();
-    bool processmove(uint32_t x, uint32_t y,Entity* e);
-    bool processmovelocal(uint32_t x, uint32_t y);
+    bool processmove(uint16_t x, uint16_t y,Entity* e);
+    bool processmovelocal(uint16_t x, uint16_t y);
     bool ishost();
     bool isprocessing();
     uint16_t getstatusforgameplay();
